pyproject.toml
[tool.poetry]
name = "job-radar"
version = "0.1.0"
description = "High-performance job aggregation engine"
authors = ["Andrew Elbaneh <onebuffalolabs@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.112.0"}
sqlmodel = "^0.0.16"
asyncpg = "^0.29.0"
alembic = "^1.13.1"
pydantic-settings = "^2.2.1"
httpx = "^0.27.0"
celery = "^5.4.0"
redis = "^5.0.8"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
ruff = "^0.3.0"
pytest-asyncio = "^0.25.0"
respx = "^0.22.0"

[tool.pytest.ini_options]
asyncio_mode = "auto"
# This setting aligns the loop scope with your fixtures to prevent "attached to different loop" errors
asyncio_default_fixture_loop_scope = "function"
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning",
    # Suppress noise from asyncpg/sqlalchemy during cleanup if tests are passing
    "ignore:.*Task attached to a different loop.*:RuntimeWarning"
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I"]